requirements-dev.txt
asgiref==3.3.4
beautifulsoup4==4.9.3
Django==3.2.2
django-bootstrap4==3.0.1
django-debug-toolbar==3.2.1
django-filter==2.4.0
djangorestframework==3.12.4
Markdown==3.3.4
python-decouple==3.4
pytz==2021.1
soupsieve==2.2.1
sqlparse==0.4.1